Group view

+++ src/main/java/com/peter/kist/repository/GroupRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 }
